Adds generator for creating configs
If no config file (.yo-rc.json) exists, you're forced into creating one
the first time that you run `yo cpui`. It allows you to override some
defaults, with an exception of separators: I just think that's more
confusing during config setup than it is helpful. It can always be
modified later on, manually.

There's also a convention I tend to use, which is instead of using the
`self` keyword, I like to name the object what it is so that the context
is immediately clear.

Conflicts:
	lib/cpui-base.js