Moves default config into json file
Since `this.src` always refers to the sub-generator's templates folder, we can at least be sure that the relative path is the same for each sub-generator. This was the best way I could think to get access to that path. Also removes the .save() line since .defaults() does that for us.