Forces config creation if none exists
This removes the config "bootstrap" generator. It was nice to have as
a choice, but if it lives as a generator it's a lot harder to invoke
from other steps in the generator process. Having it become part of
cpui-base made the most sense since this applies to every facet of
generation for this.

So now, in the init() method of each generator there's a check for
a config and a call to create one if it doesn't exist. If it does exist,
it moves on to the _create() method, which is now in every generator's
index.js.

Right now, we're only checking to see if there are no config items set,
but ideally we'd be checking to make sure that at least all the
necessary keys exist. That's a next step, but for now, this solution
works and is no longer so DRY that it chafes ;)