Ugh
This is in the middle of two possible solutions. I'm going to move the bootstrap stuff back into a generator, because it needs to be a generator as well as happen automatically. Perhaps I need to extract the common methods into a module that i can include in both places.